requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn==0.24.0

# Database
sqlalchemy==2.0.25
pymongo==4.6.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Security and Authentication
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.9

# Email validation
email-validator==2.1.1

# Optional but recommended
python-dotenv==1.0.0

# JWT
PyJWT==2.8.0